+++ src/components/common/modal/ModalVariants.ts
