+++ src/components/common/pagination/PaginationButtonVariants.ts
